backend/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
Pillow==12.0.0
python-dotenv==1.2.1
asgiref==3.11.0
sqlparse==0.5.4
pytz==2025.2
tzdata==2025.2
